% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromKey.R
\name{cdmFromKey}
\alias{cdmFromKey}
\title{Create a CDM reference from a saved database key}
\usage{
cdmFromKey(key, writePrefix = "omopbridge_")
}
\arguments{
\item{key}{Name passed to \code{\link[=setupDatabaseKey]{setupDatabaseKey()}}.}

\item{writePrefix}{Prefix used for tables written by CDMConnector.}
}
\value{
A CDM reference.
}
\description{
Create a CDM reference from a saved database key
}
